feat : libft

+++ libft/ft_striteri.c
@@ -12,9 +12,15 @@
 
 #include "libft.h"
 
+// Pass characters and indexes from the string to the function
+// Run a function for each index
+
+// Applies the function ’f’ on each character of the string passed as argument,
+// passing its index as first argument.
+// Each character is passed by address to ’f’ to be modified if necessary.
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (s[i])
